src/logion/controllers/legalofficer.controller.ts
import { OpenAPIV3 } from "openapi-types";
import { injectable } from "inversify";
import { addTag, setControllerTag, getDefaultResponses, addPathParameter } from "./doc";
import { Controller, ApiController, HttpGet, Async } from "dinoloop";
import { components } from "./components";
import { LEGAL_OFFICERS } from "./testdata";

export function fillInSpec(spec: OpenAPIV3.Document): void {
    const tagName = 'Legal Officers';
    addTag(spec, {
        name: tagName,
        description: "Retrieval and Management of Legal Officers details"
    });
    setControllerTag(spec, /^\/api\/legal-officer.*/, tagName);

    LegalOfficerController.fetchLegalOfficers(spec);
    LegalOfficerController.getLegalOfficer(spec);
}

type LegalOfficerView = components["schemas"]["LegalOfficerView"]
type FetchLegalOfficersView = components["schemas"]["FetchLegalOfficersView"]

@injectable()
@Controller('/legal-officer')
export class LegalOfficerController extends ApiController {

    constructor() {
        super();
    }

    static fetchLegalOfficers(spec: OpenAPIV3.Document) {
        const operationObject = spec.paths["/api/legal-officer"].get!;
        operationObject.summary = "Gets the list of all legal officers";
        operationObject.description = "No authentication required.";
        operationObject.responses = getDefaultResponses("FetchLegalOfficersView");
    }

    @HttpGet('')
    @Async()
    async fetchLegalOfficers(): Promise<FetchLegalOfficersView> {
        return { legalOfficers: LEGAL_OFFICERS };
    }

    static getLegalOfficer(spec: OpenAPIV3.Document) {
        const operationObject = spec.paths["/api/legal-officer/{address}"].get!;
        operationObject.summary = "Gets the details of one legal officer";
        operationObject.description = "No authentication required.";
        operationObject.responses = getDefaultResponses("LegalOfficerView");
        addPathParameter(operationObject, 'address', "The Polkadot address of the expected Legal Officer")
    }

    @HttpGet('/:address')
    @Async()
    async getLegalOfficer(address: string): Promise<LegalOfficerView> {
        const legalOfficer = LEGAL_OFFICERS.find(lo => lo.address === address);
        if (legalOfficer) {
            return legalOfficer
        }
        throw new Error(`Unknown address: ${address}`)
    }
}
